Check description for details
Added GroupInvoker helper class to simultaneously invoke a function of nodes in a group.
PopupVariableSetter has global and local key setter.
Added "adding" function in VariantWatcher for adding variables as global or local variables.
Added "copying" function in VariantWatcher, the same as "adding" but the values are displayed in the setter UI.
Added "deleting" function in VariantWatcher for removing a variable.
Added custom data for OptionButton to use in OptionValueControl.
Added separator config for adding item in PopupContextMenu.
Added edit flags in PopupVariableSetter.
Added functions to get/set local and global key setters.